SERVER-67315 Fix possible int32 overflow on call to DocumentSourceSort::createBoundedSort()
(cherry picked from commit c5b3c193d802c5618db349af4efdadbca5e59125)

+++ src/mongo/db/pipeline/pipeline_d.cpp
@@ -1409,8 +1409,9 @@ PipelineD::buildInnerQueryExecutorGeneric(const MultipleCollectionAccessor& coll
                                 sort->getSortKeyPattern(),
                                 (indexOrderedByMinTime ? DocumentSourceSort::kMin
                                                        : DocumentSourceSort::kMax),
-                                ((indexOrderedByMinTime) ? unpack->getBucketMaxSpanSeconds()
-                                                         : -unpack->getBucketMaxSpanSeconds()) *
+                                static_cast<long long>((indexOrderedByMinTime)
+                                                           ? unpack->getBucketMaxSpanSeconds()
+                                                           : -unpack->getBucketMaxSpanSeconds()) *
                                     1000,
                                 sort->getLimit(),
                                 expCtx));
